app/researcher/multi_agents/requirements.txt
langgraph
icis_researcher
langgraph-cli
python-dotenv
weasyprint
json5
loguru
